@@ -9,40 +9,40 @@
     "lint:fix": "npm run lint -- --fix"
   },
   "dependencies": {
-    "@fontsource/hind": "^4.5.10",
-    "@fontsource/open-sans": "^4.5.13",
-    "@mdi/js": "^7.0.96",
+    "@fontsource/hind": "^5.0.8",
+    "@fontsource/open-sans": "^5.0.8",
+    "@mdi/js": "^7.2.96",
     "@mdi/react": "^1.6.1",
-    "algoliasearch": "4.14.2",
-    "instantsearch.css": "7.4.5",
-    "medium-zoom": "^1.0.6",
+    "algoliasearch": "4.19.1",
+    "instantsearch.css": "8.0.0",
+    "medium-zoom": "^1.0.8",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-instantsearch-dom": "6.38.0",
+    "react-instantsearch-dom": "^6.40.4",
     "react-modal": "^3.16.1",
     "react-scripts": "5.0.1"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.19.1",
-    "@babel/preset-react": "^7.18.6",
-    "eslint": "^8.26.0",
+    "@babel/eslint-parser": "^7.22.10",
+    "@babel/preset-react": "^7.22.5",
+    "eslint": "^8.47.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-algolia": "^20.0.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-algolia": "^22.0.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsdoc": "^39.3.25",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.10",
+    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-jsdoc": "^46.4.6",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.1",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.7.1",
+    "prettier": "^3.0.1",
     "prop-types": "15.8.1",
-    "workbox-core": "^6.5.4",
-    "workbox-expiration": "^6.5.4",
-    "workbox-precaching": "^6.5.4",
-    "workbox-routing": "^6.5.4",
-    "workbox-strategies": "^6.5.4"
+    "workbox-core": "^7.0.0",
+    "workbox-expiration": "^7.0.0",
+    "workbox-precaching": "^7.0.0",
+    "workbox-routing": "^7.0.0",
+    "workbox-strategies": "^7.0.0"
   },
   "browserslist": [
     ">0.2%",
